Progress, next is to check sass
