+++ src/frontend/src/index.tsx
@@ -1,11 +1,24 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
-import App from './components/app/App';
-import './main.style.scss';
-import registerServiceWorker from './registerServiceWorker';
+// These imports are all defaults from create-react-app.
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './index.css';
+import App from './App';
+import * as serviceWorker from './serviceWorker';
 
+// The following is based on the example code at
+// https://react-redux.js.org/introduction/quick-start
+import { Provider } from 'react-redux';
+import store from './redux/store';
+
+const rootElement = document.getElementById('root');
 ReactDOM.render(
-  <App />,
-  document.getElementById('root') as HTMLElement
-);
-registerServiceWorker();
+  <Provider store={store}>
+    <App />
+  </Provider>,
+  rootElement
+)
+
+// If you want your app to work offline and load faster, you can change
+// unregister() to register() below. Note this comes with some pitfalls.
+// Learn more about service workers: https://bit.ly/CRA-PWA
+serviceWorker.unregister();
