+++ src/frontend/src/redux/reducers/index.tsx
@@ -1,0 +1,3 @@
+import { combineReducers } from "redux";
+
+export default combineReducers({ });
