+++ src/new-frontend/package.json
@@ -2,6 +2,7 @@
   "name": "record-expunge-pdx",
   "version": "0.1.0",
   "private": true,
+  "proxy": "http://localhost:5000",
   "dependencies": {
     "@types/jest": "24.0.11",
     "@types/node": "11.13.8",
