+++ src/new-frontend/public/index.html
@@ -19,9 +19,9 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
+    <title>Record Expunge</title>
   </head>
-  <body>
+  <body class="bg-gray-blue black-60">
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
     <!--
